Day 6: large performance optimisation
Big sets are slow, modifying sets is slow. Avoid that for a lot
of the path by only tracking corners for loop detection.

In theory this wastes a bit of time as we might have to trace
the loop a bit further, but it's massively outweighed by the
time saved dealing with set operations.